Restructure meta-fus layers
Move some recipes from meta-fus-bsp to meta-fus-sdk, because the bsp
layer should only contain recipes, which are absolute necessary to get
the board running.

Also move the qt6 and basler images to the dynamic layers, to avoid
dependency issues if these layers are not present.

Remove old recipes, which are not used anymore.