Use github for F&S specific gits
We used local gits for the F&S specific recipes like U-Boot and Linux.
These gits were created by the yocto download script that came with our
releases. But some customers want to be able to build the meta-fus layer
without our release scripts.
So we use the github gits by default but still add the possibility to
use the local gits by uncommenting a line, if the customer wants to.