Add tools necessary for signing fsimx93 fsimages
To sign an NXP Container we need a way to know specific offsets.
These are either found in the build logs, or by computing them. We
want to compute them ourself to not have to rely on the logs. For
this we can use a modified fsimage.sh that, additionally to list
the containers and sub images, lists the offset values.

Add NXPs Code Signing Tool to the ahab depends. Use a already
existing implementation. This allows to sign the images in ahab.bb.