recipes-bsp: Improve Dependency Decleraton to Prevent Race Conitions
Previously, the DEPENDS += ... declaration ensured that the do_patch()
task of the recipe was dependent on the do_populate_sysroot() tasks of
the specified packages. However, since we rely on files from DEPLOY_DIR,
it is essential to wait for the do_deploy task instead.

Failing to explicitly wait for do_deploy can lead to race conditions,
potentially causing build errors.

This change refines the dependency handling to accurately represent the
required task dependencies, ensuring a more robust build process.