Make Kernel signing more robust
In rare cases someone may only want to sign the bootloaders.
Modify ahab.bb, so it does not fail if no fit-Image is found.

For Yocto to build a fit-Image, a few Variables need to be set.
Instead of adding them in the local.conf every time, add them in
the machine config for fsimx93. Tha way there will always be a
fit-Image build.