Do not use meta-imx as base layer anymore
meta-imx is a proof of concept layer, which does not offer long time
support and has many quick and dirty solutions, which make it difficult
to maintain own board families.

We want to switch to the more stable and community driven meta-freescale
layer, which offers LTS.
For this, we have to make some changes to the meta-fus layer, so the
images still behave like before.

We want out ethernet devices to still be named eth0/1 for now, so we
have to disable the sytemd Predictable Network Interface Names. In
future releases we might change this.

The imx93 pxp  and cma devices need certain access rights in order to
work with weston, so we have to add them to the video user group.
It might me better to add an own group for cma access, but since weston
might not need this access anymore in future releases, we do tis the
simple way for now.

Rework fus-setup-release to do the work of imx-setup-release.