images:fus-image-qt6: fix populate_sdk
This patch will inherit populate_sdk_qt6 instead of
populate_sdk_qt6_base.

This ensures that all QT tools such as qmake and QT libraries are
available in the SDK.

Also:
QTBase is built with QT_FEATURE opengles3, opengles31 and opengles32.

OpenGLES3x is available from mesa as virtual/libgles2 and
virtual/libgles3 as a single library libGLESv2.so. libgles3 headers are
required.