[backport]fs-setup-release: ensures active shell, when sourced script failes
This is a backport of scarthgap

fus-setup-release is typically sourced.

exit 1 will close the active shell, so return instead.

Set +/-e is used so that the script is interrupted in case of errors.
It must be ensured that only the sourcing of the script is cancelled,
but not the active shell!

=> return always with 0