Add flexible chromium and spdx support
The new chromium needs a special rust version.
We don't want to use this version by default and want to keep the layers
added to the build by default as small as possible.
So we add a possibility to add chromium support with the
fus-setup-release script with --add-chromium flag.

We also wanto to be able to add spdx generation to the build.
This can be done by adding --add-spdx [organisation]