Fix kernel-module-nxp-wlan moprobe issue for fsimx8mm
The fsimx8mm cpus do not use the nxp-wlan module per default and they do
not use the btnxpuart at all.

the mxm-wifiex-filter.conf forces the load of the modules, which causes
errors if the nxp driver is not activated in mxm-wifiex.conf.

Only install the mxm-wifiex-filter.conf for boards with btnxpuart.